Create BaseStreamProcessor<T> and update ContentStreamProcessor to implement BaseStreamProcessor<PageContent>
