+++ src/UglyToad.PdfPig/Graphics/XObjectContentRecord.cs
@@ -8,20 +8,38 @@
     using Util.JetBrains.Annotations;
     using XObjects;
 
-    internal class XObjectContentRecord
+    /// <summary>
+    /// An XObject content record.
+    /// </summary>
+    public class XObjectContentRecord
     {
+        /// <summary>
+        /// The XObject type.
+        /// </summary>
         public XObjectType Type { get; }
 
+        /// <summary>
+        /// XObject stream.
+        /// </summary>
         [NotNull]
         public StreamToken Stream { get; }
 
+        /// <summary>
+        /// The applied transformation.
+        /// </summary>
         public TransformationMatrix AppliedTransformation { get; }
 
+        /// <summary>
+        /// The default rendering intent.
+        /// </summary>
         public RenderingIntent DefaultRenderingIntent { get; }
 
+        /// <summary>
+        /// The default color space.
+        /// </summary>
         public ColorSpace DefaultColorSpace { get; }
 
-        public XObjectContentRecord(XObjectType type, StreamToken stream, TransformationMatrix appliedTransformation,
+        internal XObjectContentRecord(XObjectType type, StreamToken stream, TransformationMatrix appliedTransformation,
             RenderingIntent defaultRenderingIntent,
             ColorSpace defaultColorSpace)
         {
