+++ src/UglyToad.PdfPig/PdfFonts/IFont.cs
@@ -4,20 +4,44 @@
     using System.Collections.Generic;
     using Tokens;
 
-    internal interface IFont
+    /// <summary>
+    /// Font base interface.
+    /// </summary>
+    public interface IFont
     {
+        /// <summary>
+        /// The font name.
+        /// </summary>
         NameToken Name { get; }
 
+        /// <summary>
+        /// Is the font vertical.
+        /// </summary>
         bool IsVertical { get; }
 
+        /// <summary>
+        /// The font details.
+        /// </summary>
         FontDetails Details { get; }
 
+        /// <summary>
+        /// Read the character code.
+        /// </summary>
         int ReadCharacterCode(IInputBytes bytes, out int codeLength);
 
+        /// <summary>
+        /// Try get the unicode value.
+        /// </summary>
         bool TryGetUnicode(int characterCode, out string value);
 
+        /// <summary>
+        /// Get the font bounding box.
+        /// </summary>
         CharacterBoundingBox GetBoundingBox(int characterCode);
 
+        /// <summary>
+        /// Get the font transformation matrix.
+        /// </summary>
         TransformationMatrix GetFontMatrix();
 
         /// <summary>
