+++ src/UglyToad.PdfPig/XObjects/XObjectType.cs
@@ -1,9 +1,23 @@
 ﻿namespace UglyToad.PdfPig.XObjects
 {
-    internal enum XObjectType
+    /// <summary>
+    /// XObject type.
+    /// </summary>
+    public enum XObjectType
     {
+        /// <summary>
+        /// Image.
+        /// </summary>
         Image,
+
+        /// <summary>
+        /// Form.
+        /// </summary>
         Form,
+
+        /// <summary>
+        /// PostScript.
+        /// </summary>
         PostScript
     }
 }
